Initial Commit
Comments are previous experiments. Parents gets you trapped in a recursive loop. If I felt like it, I could add a list of AT_URIs to the clout_dict and compare against (that way in case someone replied with a quote, it could count the parents) but 1. I don't even know if that's what I want as output and 2. That seems like a lot of extra overhead for a pretty niche case (lol, lmao, overhead in python).